JVMCBC-427: Explicitly only look for IPv4 Addrs
Motivation
----------
If DNS AAAA records are checked before A records and
IPv6 is forced, it can be the case that the SDK tries
to connect to an Couchbase Server via IPv6 which the
SDK supports but the server doesn't.

Modifications
-------------
The NetworkAddress class now explicitly iterates through
all the found InetAddress instances and looks for one that
maps to IPv4. If none is found, an exception is raised.

Result
------
Better fallback support to IPv4 if the JVM/DNS setup enforces
IPv6.

Change-Id: I7a75413c09e3511bded77f17b816fdd5ac7f6015
Reviewed-on: http://review.couchbase.org/78393
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>