Use admin properties for test authentication.
Motivation
----------
With Couchbase Server 5.0 and RBAC coming, it isn't enough to just
use bucket credentials, also users need to be in place (especially
when working with buckets without passwords!).

Modifications
-------------
Because the admin user has all credentials and is always available,
it makes sense to just authenticate with that user in order to run
the test suite completely.

Result
------
The test suite doesn't fail on the int test with invalid credentials.

Change-Id: Ib012a050428eb8451e378369ddbcf533e4f830d4
Reviewed-on: http://review.couchbase.org/78711
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>