Make test suite compatible with RBAC.
Motivation
----------
The test suite got modified to work with 5.0 but
then it didn't work with older releases. The
test suite needs to run transparently against
all supported releases properly.

Modifications
-------------
A helper method has been introduced which, before
bootstrapping, loads the config and grabs the
cluster node version.

If its >= 5, then the username and password are
set to admin credentials which are guaranteed
to work in the RBAC scenario. Otherwise, they
stay the same as before (bucket level).

Result
------
The test suite now runs against pre-5.0 and post
5.0 releases.

Change-Id: Idd98458f9f673c971f8670ff3b8e48bb3e0af115
Reviewed-on: http://review.couchbase.org/78828
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>