JVMCBC-431 Add support for insertDocument and accessDeleted in subdoc doc flags
Motivation
----------
Include support for insert document and access deleted in subdocument
document flags.

Changes
-------
The add document flag is set for inserting new document in a subdocument
mutation request and the access deleted flag is set for accessing deleted
documents on subdocument requests. These options are set as part of the
request.

Results
-------
The two doc flags work as expected with 5.0 server release.

Change-Id: I7e228a9e0cef3418c6d7a717ff77bfbf791c94b8
Reviewed-on: http://review.couchbase.org/79445
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>