JVMCBC-433: Add X.509 client side cert auth support.
Motivation
----------
Couchbase Server 5.0 adds support for authenticating through
client side certificates when using SSL/TLS, so the client
needs to support it as well.

Modifications
-------------
Most of the work is done for us in Netty/JVM, but the important
part is that we need to skip SASL auth (since that is implicit
with the certificate) when enabled.

To expose it nicely to the user a setter has been added to the
CoreEnvironment and if set to true the endpoint is skipping
the SASL handlers in the pipeline during setup.

To avoid misconfiguration, the environment also checks the
invariant to make sure SSL is enabled when client side cert
support is.

Result
------
Support for X.509 client side cert authentication.

Change-Id: Icd225e452bcab07fe2b546e154c9367d5b3c4b8e
Reviewed-on: http://review.couchbase.org/79461
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>