JVMCBC-434: Add low-level support for "Enhanced Error Messages".
Motivation
----------
In order to support the new 5.0 "enhanced error message" feature where
the server sends more infos in the error case to the SDK, the client
needs to be able to parse those conditionally anf if so make it
available to the upper layers.

Modifications
-------------
Because only ResponseStatus enum has been supported, a new getter and
setter have been added to the CouchbaseResponse (statusDetails) which,
if set and not null, contains the additional reference and context.

The KV handler has been modified to conditionally decode this information
if needed and set it on the response if not null.

Result
------
Support for enhanced error messages.

Change-Id: Ib1218b11b9e378998ee13b38740ec59fb25c6689
Reviewed-on: http://review.couchbase.org/79472
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>