Disable XERROR in tests by default.
Motivation
----------
Right now the XERROR tests are not universal and
stable enough to make them always pass against all
kinds of configurations.

This changeset disables it by default and skips tests
that depend on it.

This makes the test suite green again, a separate
effort needs to be taken to make the kv error map
changes more stable to enable them by default again.

Change-Id: Ie6b2e6d1fe399735e977b01a4c55a70a9636941f
Reviewed-on: http://review.couchbase.org/80511
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>