JVMCBC-439: Fix prepared statement dispatching with node hostname
Motivation
----------
This is a regression. When the NetworkAddress has been introduced
this codepath in the QueryLocator has not been modified and now
the equality comparison always fails between InetAddr and NetworkAddress
- thanks java for not failng to compile ;-)

Modifications
-------------
Convert to a NetworkAddress on the fly and as a result handle the
equality comparison properly, fixing the regression and allowing the
Prelocated Requests to dispatch properly.

Result
------
Regression fixed.

Change-Id: I3007e25648ee34c0affd0beccab13d1a8c0522df
Reviewed-on: http://review.couchbase.org/80574
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>