JVMCBC-441: Add a CoreSendHook
Motivation
----------
Right now it is not possible to hook in before we "send" the request
into the ringbuffer, allowing to introspect the request and also
to potentially alternate the observable returned to the upper layers.

Modifications
-------------
This changeset makes it possible to hook into the send method, configurable
via a hook through the environment. The tuple returned allows to either
pass through or modify the original request and/or the original response.

Result
------
It is now possible to optionally hook the send method before the request
is put into the ringbuffer.

Change-Id: I7501f1a3cc93275e26e7644248b0261feecb09aa
Reviewed-on: http://review.couchbase.org/81108
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>