JVMCBC-442: Once dispatched, store the destination in the request.
Motivation
----------
For traceability reasons, it makes sense that once a request is
dispatched to a socket we actually also store where it got sent.

This helps when someone is looking at the request later on to
figure out where it went and can use it for monitoring and/or
retry capabilities.

Modifications
-------------
The CouchbaseRequest and its immediate surroundings are modified
to contain an additional field which is filled by the endpoint.

Note that, by design, when the operation gets rescheduled after
being dispatched the location might change again, so it always
contains the last dispatch try. If its null it hasn't been dispatched
yet which is another good source of information.

Result
------
Better observability on where a request got dispatched into.

Change-Id: I4de29c6e87b5df0b66dd91eb1e461e2a5fb6f077
Reviewed-on: http://review.couchbase.org/81247
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>