WIP for JCBC-1099 : N1ql query result parsing has poor performance for big document
This is a WIP that should be heavily reviewed / updated by a couchbase
java sdk ninja

Motivation
----------
Simple bucket with only 2 (heavy) documents.
Simple query select b.* FROM thebucket b
Execute the query in the couchbase admin : 120 ms
the same query with the java client : 12 seconds
the thread doing the computation is using a core at 100%
the document is huge and badly designed but that's not the point
the performances should be nearly the same between java and the admin
console.
After investigation the performance bottleneck is in
QueryHandler#parseQueryRows >
ByteBufJsonHelper#findSectionClosingPosition
Looks like data in being appended to a buffer but the scan with
forEachByte is always done from the beginning of the buffer.

Modifications
-------------
In QueryHandler#parseQueryRows reuse the ClosingPositionBufProcessor and
restart the scan when when it ends at the last run with 
int forEachByte(int index, int length, ByteBufProcessor processor)

Result
------
It brings the performance on par with the query launched from the
console 12s -> ~200ms