JVMCBC-443: Remove DCP from Core-IO
Motivation
----------
DCP has been factored out into its own project (which has core-io
as a dependency) and we've deprecated it couple months ago.

Modifications
-------------
Now that 1.5.0 is around the corner DCP is removed from core-io
and only lives in a standalone project where it is maintained.

Result
------
Getting rid of unmaintianed and unused code in core-io making it
leaner.

Change-Id: I51a2652eb530f46d9fe57b346b399ea341890f09
Reviewed-on: http://review.couchbase.org/81977
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>