JVMCBC-444: Enforce config poll floor at 50ms.
Motivation
----------
According to the fast failover RFC, the config poll
floor should be enforced at 50ms, so polling should
not occur more frequently than that.

Modifications
-------------
This changeset changes the environmental setting so
it can be set as low as 50ms but it also adjusts the
carrier refresher (which is responsible for doing the
polling) to ignore all refresh attempts below 50ms so
those triggered from other sources do not happen
higher than the given interval.

Result
------
Config poll floor at 50ms enforced as specified in the
fast failover RFC.

Change-Id: Icbc6d3f3e480f5f5f5fadf2ed7daad27d922ef46
Reviewed-on: http://review.couchbase.org/81986
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>