JVMCBC-445: Use an iterator on pool flush to avoid IOOB.
Motivation
----------
In the PooledService flush implementation, we are iterating
the CopyOnWriteArrayList in a simple for loop with direct
index access - this works as long as there is no concurrent
access on the underlying structure.

If this happens, a ArrayIndexOutOfBoundsException is
possible.

Modifications
-------------
To avoid the ArrayIndexOutOfBoundsException, the COWArrayList
provides a stable iterator() which copies the original list
before iterating, preventing the race condition from
happening in the first place.

The changeset uses the iterator approach rather than the
manual indexing.

Result
------
No more ArrayIndexOutOfBoundsException when iterating and
modifying the endpoint list simultaneously.

Change-Id: Ie123d6df31cc00db514630e9b1be0a41b14877f7
Reviewed-on: http://review.couchbase.org/82036
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/84186
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>