Allow to mock the ServiceFactory for Node tests.
By changing the CouchbaseNode instance to allow mocking
the ServiceFactory, the test case can be modified to pass
in a custom one which avoids going to the network at all.

This makes sure the unit tests don't actually try to open
a socket against port 0, leading to failures on jenkins.

Change-Id: Iab049447605b1f7510ad5f6a6ef299557cda1416
Reviewed-on: http://review.couchbase.org/82046
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>