JVMCBC-449: Sanity check delay upper and lower bounds.
Motivation
----------
The lower and upper boundaries can be easy to mess up and a user
was running into troubles because of this. We need to sanity
check on creation, making it easier to spot user errors.

Modifications
-------------
For linear and exponential delay, the upper and lower boundaries
are asserted and an exceptio is raised immediately if the lower
boundary is larger than the upper one.

Result
------
Immediate feedback on parameter misuse/switch without knowing
whats going on otherwise.

Change-Id: If28a71c9fddf49c98fad32e2213e4a39ab87e015
Reviewed-on: http://review.couchbase.org/82199
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>