JVMCBC-451: Add support for Subdocument Get Count
Motivation
----------
In Couchbase Server 5.0, a new lookup command (get count) is introduced
which allows to fetch the number of items for a given path in a document.

Modifications
-------------
This changeset adds the new command to the Handler opcodes and makes sure
that the appropriate single and multi path messages are setup so that
they can be encoded correctly.

Result
------
core-io support for subdocument get count.

Change-Id: I848478a0606330648dc033b696086dcc1984d889
Reviewed-on: http://review.couchbase.org/82441
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>