JCBC-1096: Explicitly handle auth errors on observe.
Motivation
----------
With Server 5.0 and RBAC it is now possible that a user only has
write access but not read access, leading to the situation where
a PersistTo or ReplicateTo might fail after correctly storing
the doc but then the user is not allowed to use the observe
command.

Modifications
-------------
This change explicitly checks for authentication errors and
supplies a helpful message, including the enhanced error
messages back to the user.

Result
------
If the user is performing an op and it now fails because
of an auth issue, the error message is now clear.

Change-Id: I252afa4dcad50078474285d13c0c38269b17f46d
Reviewed-on: http://review.couchbase.org/82623
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/84188
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>