JVMCBC-435: Limit config poll floor only on a per-bucket basis.
Motivation
----------
The config poll floor added in JVMCBC-415 is causing a regression
with properly loading and applying bucket configurations when more
than one bucket is opened at the same time.

Modifications
-------------
The change applied in JVMCBC-415 is applying a refresher-global
poll floor for new bucket configurations. This is problematic
because config refresh attempts for _different buckets_ that are
quickly sent one after another are dropped if below the configured
min floor interval.

Instead, the floor should be in place but it needs to be checked
on a per-bucket basis, giving each bucket a fair chance to grab
a new configuration.

This also remedies the side effect of all this that multiple
bucket configs can end up with varying node list sizes which
in turn results in keeping connections open to nodes which
are not part of the cluster anymore (consider a hard failover).

Result
------
By tracking the poll floor on a per-bucket basis, each bucket
is given a fair chance to grab a new configuration, resulting
in prompt config updates even if more than one bucket is open.

Change-Id: I19dca0da8e541a48afc96c06378a420f93b00ca2
Reviewed-on: http://review.couchbase.org/82917
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>