JVMCBC-457: Make new query parser the default.
Motivation
----------
The new generic streaming query parser is more reliable and
makes it possible to avoid one-off response parser that are
fragile.

Modifications
-------------
This one has been in the making for a while but behind a feature
flag. With the upcoming minore release its time to promote it to
active by default, but it can always be disabled as a workaround
if any bugs should surface that haven't been covered.

Result
------
The new response parser is enabled by default!

Change-Id: I88e53bc3055da8111bde8ba44478698775c6bffa
Reviewed-on: http://review.couchbase.org/82948
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>