JVMCBC-456: current request needs to be set on response.
Motivation
----------
The old query handler did set the currentRequest on the
GenericQueryResponse, but the new parser didn't even if the
response expects it to.

Modifications
-------------
This changeset adds a test to assert for non-nullness of the
request and makes sure it gets set properly by the new parser.

Result
------
The query response has the original request properly set by
the new parser.

Change-Id: I09ec17e2cc841342d339a620aa94441c7c7ffa21
Reviewed-on: http://review.couchbase.org/82951
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>