JVMCBC-456: Remove unneeded observeOn and fix backpressure.
Motivation
----------
The current parser has two pieces of code which are not working
as intended. Both the observeOn and onBackpressureBuffer calls
actually return the modified object and as a result are ignored
right now.

Modifications
-------------
As it turns out after investigation, the observeOn call is actually
not needed because it doesn't matter from where the subjects are
completed, what matters is where they are consumed from - and this
is handled in a different place.

As a result, this changeset gets rid of the observeOn calls. Since
it has never had any effect the code keeps working as before with
that removal, its just less ambiguous.

The other change in place puts the onBackpressureBuffer() call in
a place where it is actually used, and this is right when the
response is created. Now the observable returned to the user includes
the onBackpressureBuffer() semantics properly.

Result
------
After this changeset is applied, the observeOn semantics are
clarified and the onBackpressureBuffer method is properly applied
to the response.

Change-Id: I6935ae134a7fd7282b95b7bccb35696134a11989
Reviewed-on: http://review.couchbase.org/82961
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>