JVMCBC-456: Use Deque instead of Stack
Motivation
----------
Stack is an older and more or less deprecated collections class
in java, and the JDK recommends moving to a Deque implementation
instead. This also has the benefit of not having to use a
synchronized datastructure where we don't need one, paying an
overhead possibly.

See https://docs.oracle.com/javase/7/docs/api/java/util/Deque.html
for more details.

Modifications
-------------
One can use the Deque as a drop-in-replacement for the Stack since
it provides all the necessary methods. Test pass.

Results
-------
Use a more modern, unsynchronized datastructure instead of the old
Stack one.

Change-Id: I455509c53e628afda08322fdcb84c60ec2689c24
Reviewed-on: http://review.couchbase.org/82976
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>