JVMCBC-458: Discover Analytics service through the server config.
Motivation
----------
The analytics service is now distributed (dp3+) from the server
config in specially enabled builds. If one of those is used we
should pick up the service right from the server config like
with all the other services.

Modifications
-------------
The port info mapping is updated to take the "cbas" and "cbasSSL"
port mappings into account. Note that this now would override
the ports set by the manual analytics enabled properties, but
they are intended to go away anyways once analytics hits GA.

Result
------
Analytics service is now automatically discovered when the
service is enabled on the server.

Change-Id: Ia0e7d4659811858a2ce11716ce67c467ed2d112a
Reviewed-on: http://review.couchbase.org/83064
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>