JVMCBC-461: Add basic health check API support.
Motivation
----------
To expose a simple "health check" api at the upper layers
that provide quick insight into the sockets connected in
the SDK, the core-io layer is responsible for assembling
all this information and presenting it up the stack in an
easy to use fashion.

Modifications
-------------
This changset does a couple of things:

 - it adds the Reqeusts and Responses for the health check
 - it propagates the health check from the way up down to the endpoint
   and collects/merges it at each individual layer
 - in the endpoint, it assembles the EndpointHealth as a final instance
   and passes it back up the stack.

So in a tree fashion the state for all the currently known endpoints
is aggregated and in the end passed back to the caller.

Aside from the basic socket information it also accesses the last known
activity (from any request) and piggybacks on the health checks to record
the latency for the last known health check.

Result
------
The basic groundwork for the health check API is in place and can
be used.

Change-Id: I2ebef38bb79893bdef519c1a2233f028b0991bbc
Reviewed-on: http://review.couchbase.org/83079
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>