JVMCBC-466: Cancel KeepAlive future when channel is removed.
Motivation
----------
When "continuous keepalive" is enabled, which it is by default
now, then when a socket gets closed the keepalive still keeps
going (or at least tries to), since the scheduled task is not
cancelled.

Modifications
-------------
This changeset makes sure that when the handler is removed, the
future which coordinates the timed execution gets cancelled and
as a result stops the continuous execution attempts.

Result
------
When a node is properly removed the keepalive doesn't keep on
going and potentially leaking sockets.

Change-Id: I9ec94c7897f4d021a078e94d5564648c8e8bd499
Reviewed-on: http://review.couchbase.org/83879
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-on: http://review.couchbase.org/84192