JVMCBC-467: don't perform notifyChannelInactive after disconnect.
Motivation
----------
When an explicit disconnect is performed (so an endpoint gets removed
on purpose), it doesn't make sense to log the INFO message, or perform
any other operations in this method at all.

Modifications
-------------
This changeset in the beginning explicitly makes sure that the method
quickly returns when we are disconnecting explicitly.

Note that normally the side effects of this method won't be executed
anyways since the state is DISCONNECTED, but to prevent any side
effects or unintended consequences we short-circuit it right away
at the beginning.

Result
------
No side effects happening on orderly disconnect, but the main impact
is that the misleading message of "Got notified from Channel as inactive,
 attempting reconnect." is not shown when the endpoint is explicitly
disconnected (like during a rebalance out).

Change-Id: Ia574b9fb38e89e51520700685e0b077f1c71a119
Reviewed-on: http://review.couchbase.org/83880
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-on: http://review.couchbase.org/84193