JVMCBC-468: Allow to optionally specify a different TrustStore
Motivation
----------
In certain deployments it can make sense to split up the KeyManager and
TrustManager certificate management in different files / keystores.

Modifications
-------------
This changeset allows to optionally provide a different keystore for
the trust manager, if its not provided it just falls back to the
global keystore used, as before.

Note that to be semantically correct if only the truststore is provided,
it is also used for both the key store and the trust store since that's
most of the time what people want. Also, javadoc has been updated
and improved to represent the semantics better.

Result
------
More flexible certificate management with the SDK if different
keystores should be used for the key and trust managers

Change-Id: Icb025b867a43fb535fe932c7407c902078f771f7
Reviewed-on: http://review.couchbase.org/84200
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>