JVMCBC-470: Allow to configure the configPollFloorInterval
Motivation
----------
As part of the Fast-Failover RFC it was agreed to that the
config poll floor interval needs to be configurable as well.

Right now this property is fixed in the carrier refresher,
but it should also be exposed through the environment so
it can be configured accordingly.

Modifications
-------------
This changeset adds the corresponding environment builder
options and then makes use of them in the carrier refresher,
getting rid of the static value.

Result
------
The config poll floor interval is now configurable
through the environment.

Change-Id: I37b3cc1188a807f5a70440a4fd716131da144540
Reviewed-on: http://review.couchbase.org/84747
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>