JVMCBC-471: EndpointHealth needs to be resilient to null on addrs.
Motivation
----------
It is possible that when asked for EndpointHealth the endpoints
are setup but not connected yet, so the channel is still null
and as a result so are the local and remote addrs inferred.

Modifications
-------------
This is a legit situation, so the code simply needs to cope
with the possibility of being null and should properly react
with safety measures.

Result
------
The code now handles the situation outlined above. Regression
test added for good measure.

Change-Id: Ia0328095812fa0463a777a84913b267b32e46de7
Reviewed-on: http://review.couchbase.org/85008
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>