JVMCBC-472: Implement IPv6 bracket syntax support for ConnectionString
Motivation
----------
In order to add IPv6 support, it needs to be possible to pass IPv6 addr
literals when bootstrapping, which is currently broken because explicit
support for the IPv6 syntax hasn't been added.

Modifications
-------------
This changeset adds support (and tests) for the bracketed syntax which
is required in URLs as per https://www.ietf.org/rfc/rfc2732.txt . We
added support for both with and without ports, although they are not
really used at this point (but could be in the future).

Result
------
ConnectionString support for IPv6 bracketed syntax literals.

Change-Id: I728db10381c795db8fd861c2e3a402a6515293f7
Reviewed-on: http://review.couchbase.org/85183
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>