JVMCBC-475: Support NetworkAddress with IPv6
Motivation
----------
Right now we ignore IPv6 addresses on purpose since
previous couchbase releses didn't support IPv6. Now
with 5.1 coming up, we need to re-add support for
it but still allow the mode to run where the older
releases don't support it.

Modifications
-------------
If a hostname is passed in as input instead of a literal
IP address, then the class relies on
{@link InetAddress#getAllByName(String)} to perform a
lookup. By default the JVM prefers IPv4 for this lookup in
dual stack environments, unless the system property
"java.net.preferIPv6Addresses" is set to true on startup.
In this case, IPv6 is preferred and automatically selected.

Note that there is one specific case where if IPv6 is preferred
in dual stack and you are running Couchbase Server pre 5.1 which
does not support IPv6, then you must set the system property
"com.couchbase.forceIPv4" to true so that the firstly returned
IPv6 address is ignored and we are looking for an IPv4 address to be
resolved as well. Obviously, in a single IPv6 stack environment
working with older Couchbase Server releases won't work.

Result
------
Re-enable transparent IPv6 support.

Change-Id: Ib45e21828fdaa2c6630f8a275a31c3510ea06e66
Reviewed-on: http://review.couchbase.org/85439
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>