JVMCBC-477: Add SDK compatible memcached hashing.
Motivation
----------
Current memcached bucket ketama hashing strategies are
not compatible with libcouchbase and go sdk derivates,
since the way the key is hashed is a bit different.

Modifications
-------------
This changeset adds a new strategy to provide
lcb-compatible hashing and as a result it can be used
alongside lcb-built sdks.

The NodeInfo had to be adapted a bit to store and
expose the raw host string exposed from the config
in order to guarantee determinism in how the key
is hashed.

Result
------
libcouchbase-compatible hashing support.

Change-Id: I9d20479773f868ca490c8de73fb466317db43b66
Reviewed-on: http://review.couchbase.org/85685
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>