JVMCBC-473: Add forceSaslPlain auth method for LDAP compliance
Motivation
----------
If users are created on the server through LDAP, we can't use
any other mechanism than PLAIN to perform the authentication.

Modifications
-------------
This changeset introduces, as simple as possible, an environment
setting which allows to force the PLAIN authentication mode which
can be used to work together with LDAP users.

Note that we still ask for the sasl mechs list, we just throw
it away and force the use of PLAIN. We could not send the mechs
list in the first place, but testing and changing would be higher
risk than sticking with the already battle tested auth flow.

Result
------
LDAP user support can be enabled when forcing PLAIN KV
authentication.

Change-Id: Ica9617931611c0f4c4aa0ba024ce6f6dea9d9db6
Reviewed-on: http://review.couchbase.org/85896
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>