JVMCBC-479: Add completion methods for observable.
Motivation
----------
Right now throughout the library we are writing into the
observable from all kinds of places, but this ideally should
be centralized.

Modifications
-------------
This changeset lays the foundation for future refactoring by
first providing the encapsulation methods to either fail
or complete the underlying subject.

Result
------
Lays the groundwork for future refactoring around encapsulation.

Change-Id: Iea517f4272e4d3f9bd5376b141f479a8d610d164
Reviewed-on: http://review.couchbase.org/85901
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>