JVMCBC-481: Optimize config rev pre-check where possible.
Motivation
----------
It has been found that while creating a heap object out
of a json file is normally a quick operation, as part
of creating the config we are also trying to resolve
hostnames which in turn can lead to actual (blocking)
IO if the lookup happens not to be cached by the
JVM.

Modifications
-------------
While a separate changeset needs to be developed which
moves the decoding/resolution to a dedicated thread, it
makes sense to pre-check the raw config if the rev number
is actually greater than the one we have. We can do this
without loading the full config and just extracting the
rev.

If the rev is not newer, drop the raw config right away
and not even bother creating a full blown one.

The code has been restructured a little so that on the
refresh-path the config is not parsed in multiple places
and the optimization logic is consolidated. Since the
type changes, the tests had to be adapted as well.

Result
------
Quicker and definitely without IO pre-check if the config
is actually one we need to fully parse.

Change-Id: Ia7a02c227b131edb3861271f0a7819705b6a1f50
Reviewed-on: http://review.couchbase.org/86239
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>