JVMCBC-475: Force IPv4 over IPv6 for now.
Motivation
----------
With all the IPv6 support merged in, there is a
possibility that users relying on the srict
IPv4 support are seing backwards incompatible
behavior.

Modifications
-------------
In order to make the change non-intrusive in
the current bugfix release cycle, the system
property is turned off, but can be turned
on by users who want to test out the IPv6
support.

In the next minor release, this switch will
be flipped on by default and properly release
noted.

The tests also have guards to make sure they
don't fail if IPv6 is not enabled.

Result
------
All the functionality merged for going forward,
but no backwards incompatible changes on the
current bugfix branch.

Change-Id: If9820ee8df9b3d950015e0dc6309646f447f24e9
Reviewed-on: http://review.couchbase.org/86439
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>