JVMCBC-483 Accept any Object as a redactable argument, not just String
Motivation
----------
To log a redactable argument that is not a String, it was
necessary to convert it to a String first. This adds noise,
and might be wasted effort if the log message is suppressed.

Modifications
-------------
RedactableArgument.user(), meta(), and system() now accept Object
instead of String. RedactableArgument.toString() converts the object
to a String in a null-safe way.

The unit test is modified to assert the null-safety.

Result
------
No longer need to call toString() when creating RedactableArguments
from arbitrary objects.

Change-Id: Idadabb24cd7245cb84aca6bd1815163137b9d287
Reviewed-on: http://review.couchbase.org/86472
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>